# internal/swagger/swagger.json
{
  "schemes": ["http", "https"],
  "swagger": "2.0",
  "info": {
    "description": "基于Gin、Gorm、Viper、Wire、Cobra的HTTP快速开发框架 API 文档",
    "title": "EchoHub API 文档",
    "contact": {
      "name": "EchoHub Team",
      "url": "https://github.com/HoronLee/EchoHub",
      "email": "support@echohub.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
    "/helloworld": {
      "post": {
        "description": "创建一个新的HelloWorld消息并返回系统信息",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["HelloWorld"],
        "summary": "创建HelloWorld消息",
        "parameters": [
          {
            "description": "HelloWorld创建请求参数",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/helloworld.CreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建成功，返回消息和系统信息",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/helloworld.CreateResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误或创建失败",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/user/delete": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "删除当前登录的用户账户",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["用户管理"],
        "summary": "删除用户",
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "用户未认证或删除失败",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "401": {
            "description": "用户未认证",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "用户身份验证并获取访问令牌",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["用户管理"],
        "summary": "用户登录",
        "parameters": [
          {
            "description": "登录请求参数",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登录成功，返回JWT令牌",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/user.LoginResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误或登录失败",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "description": "创建新用户账户",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["用户管理"],
        "summary": "用户注册",
        "parameters": [
          {
            "description": "注册请求参数",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user.RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "注册成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误或注册失败",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "helloworld.CreateRequest": {
      "type": "object",
      "required": ["message"],
      "properties": {
        "message": {
          "type": "string",
          "maxLength": 20,
          "minLength": 1,
          "example": "Hello, World!"
        }
      }
    },
    "helloworld.CreateResponse": {
      "type": "object",
      "properties": {
        "database": {
          "type": "string",
          "example": "SQLite"
        },
        "message": {
          "type": "string",
          "example": "Hello, World!"
        },
        "version": {
          "type": "string",
          "example": "1.0.0"
        }
      }
    },
    "response.Response": {
      "type": "object",
      "properties": {
        "code": {
          "description": "Code 状态码，非0时表示自定义HTTP业务状态码",
          "type": "integer",
          "example": 0
        },
        "data": {
          "description": "Data 响应数据，具体内容因接口而异"
        },
        "msg": {
          "description": "Msg 返回信息，通常是状态描述",
          "type": "string",
          "example": "success"
        }
      }
    },
    "user.LoginRequest": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "password": {
          "type": "string",
          "example": "password123"
        },
        "username": {
          "type": "string",
          "example": "john_doe"
        }
      }
    },
    "user.LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        }
      }
    },
    "user.RegisterRequest": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "password": {
          "type": "string",
          "minLength": 6,
          "example": "password123"
        },
        "username": {
          "type": "string",
          "maxLength": 50,
          "minLength": 3,
          "example": "john_doe"
        }
      }
    }
  }
}